Move GetStyles test to LocalGeoServerOnlineTest
